Add option --google-takeout-label-priority
Extend the option --google-takeout-first-label to select a label that
makes most sense. Default is "Trash,Inbox,Sent".

Furtermore "Archived" is only selected, if no other label matches - so
it is a hardcoded low-priority-label.

When uploading a message, the destination folders are also printed to
stdout.

Lines with changes are reformated.

Update README.md with current output of "--help"